producer_app/requirements.txt
kafka-python
selenium
configparser
loguru
requests
beautifulsoup4